add export-vcenter-vms and update readme

+++ batch-enablement/arcvmware-batch-enablement.ps1
@@ -4,10 +4,7 @@ This is a helper script for enabling VMs in a vCenter in batch. The script will 
   vmware-batch.log - log file
   all-deployments-<timestamp>.txt - list of Azure portal links to all deployments created
   vmw-dep-<timestamp>-<batch>.json - ARM deployment files
-
-In a single ARM deployment, the script can enable up to 200 VMs if guest management is enabled, else 400 VMs. If there are more than 200 VMs (or 400 VMs if guest management is not enabled), the script will create multiple ARM deployments.
-
-**There are some lines which are marked with NOTE. Please read the comments and modify the script accordingly. It is recommended to run the script with the DryRun switch first to ensure that the deployments are created as expected.**
+  vmw-dep-summary.csv - summary of the VMs enabled
 
 Before running this script, please install az cli and the connectedvmware extension.
 az extension add --name connectedvmware
@@ -20,7 +17,7 @@ az login --service-principal --username <clientId> --password <clientSecret> --t
 
 Following is a sample powershell script to run the script as a cronjob:
 
-$action = New-ScheduledTaskAction -Execute 'powershell.exe' -Argument '-File "C:\Path\To\vmware-batch-enable.ps1" -VCenterId "/subscriptions/12345678-1234-1234-1234-1234567890ab/resourceGroups/contoso-rg/providers/Microsoft.ConnectedVMwarevSphere/vcenters/contoso-vcenter" -EnableGuestManagement -VMCountPerDeployment 3 -DryRun' # Adjust the parameters as needed
+$action = New-ScheduledTaskAction -Execute 'powershell.exe' -Argument '-File "C:\Path\To\vmware-batch-enable.ps1" -VCenterId "/subscriptions/12345678-1234-1234-1234-1234567890ab/resourceGroups/contoso-rg/providers/Microsoft.ConnectedVMwarevSphere/vcenters/contoso-vcenter" -EnableGuestManagement -Execute' # Adjust the parameters as needed
 $trigger = New-ScheduledTaskTrigger -Daily -At 3am  # Adjust the schedule as needed
 
 Register-ScheduledTask -Action $action -Trigger $trigger -TaskName "EnableVMs"
@@ -31,22 +28,28 @@ Unregister-ScheduledTask -TaskName "EnableVMs"
 .PARAMETER VCenterId
 The ARM ID of the vCenter where the VMs are located. For example: /subscriptions/12345678-1234-1234-1234-1234567890ab/resourceGroups/contoso-rg/providers/Microsoft.ConnectedVMwarevSphere/vcenters/contoso-vcenter
 
+.PARAMETER VMInventoryFile
+The path to the VM Inventory file. This file should be generated using the export-vcenter-vms.ps1 script, and filtered as needed. The file can be in CSV or JSON format. The format will be auto-detected using the file extension. All the VMs in the file which have VMware Tools running will be enabled.
+
 .PARAMETER EnableGuestManagement
 If this switch is specified, the script will enable guest management on the VMs. If not specified, guest management will not be enabled.
 
-.PARAMETER VMCountPerDeployment
-The number of VMs to enable per ARM deployment. The maximum value is 200 if guest management is enabled, else 400.
+.PARAMETER VMCredential
+The credentials to be used for enabling guest management on the VMs. If not specified, the script will prompt for the credentials.
 
-.PARAMETER DryRun
-If this switch is specified, the script will only create the ARM deployment files. Else, the script will also deploy the ARM deployments.
+.PARAMETER Execute
+If this switch is specified, the script will deploy the created ARM templates. If not specified, the script will only create the ARM templates and provide the summary.
 
 #>
 param(
   [Parameter(Mandatory=$true)]
   [string]$VCenterId,
+  [Parameter(Mandatory=$true)]
+  [string]$VMInventoryFile,
   [switch]$EnableGuestManagement,
   [int]$VMCountPerDeployment,
-  [switch]$DryRun
+  [PSCredential]$VMCredential,
+  [switch]$Execute
 )
 
 $logFile = Join-Path $PSScriptRoot -ChildPath "vmware-batch.log"
@@ -61,6 +64,8 @@ Write-Host "Setting the TLS Protocol for the current session to TLS 1.3 if suppo
 try { [Net.ServicePointManager]::SecurityProtocol = [Net.ServicePointManager]::SecurityProtocol -bor 12288 } catch {}
 
 $VCenterIdFormat = "/subscriptions/12345678-1234-1234-1234-1234567890ab/resourceGroups/contoso-rg/providers/Microsoft.ConnectedVMwarevSphere/vcenters/contoso-vcenter"
+
+$VMWARE_RP_NAMESPACE = "Microsoft.ConnectedVMwarevSphere"
 
 function Get-TimeStamp {
   return (Get-Date).ToUniversalTime().ToString("[yyyy-MM-ddTHH:mm:ss.fffZ]")
@@ -77,6 +82,14 @@ function LogText {
   )
   Write-Host "$(Get-TimeStamp) $Text"
   Add-Content -Path $logFile -Value "$(Get-TimeStamp) $Text"
+}
+
+function LogDebug {
+  param(
+    [Parameter(Mandatory=$true)]
+    [string]$Text
+  )
+  Add-Content -Path $logFile -Value "$(Get-TimeStamp) [Debug] $Text"
 }
 
 function LogError {
@@ -106,106 +119,99 @@ function Get-ARMPartsFromID($id) {
 #Region: ARM Template
 
 # ARM Template part for VM Creation
-$VMtpl = @'
-{
-  "type": "Microsoft.Resources/deployments",
-  "apiVersion": "2021-04-01",
-  "name": "{{vmName}}-vmcreation",
-  "properties": {
-    "mode": "Incremental",
-    "template": {
-      "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#",
-      "contentVersion": "1.0.0.0",
-      "resources": [
-        {
-          "type": "Microsoft.HybridCompute/machines",
-          "apiVersion": "2023-03-15-preview",
-          "name": "{{vmName}}",
-          "kind": "VMware",
-          "location": "{{location}}",
-          "properties": {}
-        },
-        {
-          "type": "Microsoft.ConnectedVMwarevSphere/VirtualMachineInstances",
-          "apiVersion": "2023-03-01-preview",
-          "name": "default",
-          "scope": "[concat('Microsoft.HybridCompute/machines', '/', '{{vmName}}')]",
-          "properties": {
-            "infrastructureProfile": {
-              "inventoryItemId": "{{vCenterId}}/InventoryItems/{{moRefId}}"
-            }
-          },
-          "extendedLocation": {
-            "type": "CustomLocation",
-            "name": "{{customLocationId}}"
-          },
-          "dependsOn": [
-            "[resourceId('Microsoft.HybridCompute/machines','{{vmName}}')]"
-          ]
+$VMtpl = @{
+  type       = "Microsoft.Resources/deployments"
+  apiVersion = "2021-04-01"
+  name       = "{{vmName}}-vmcreation"
+  properties = @{
+    mode     = "Incremental"
+    template = @{
+      '$schema'      = "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#"
+      contentVersion = "1.0.0.0"
+      resources      = @(
+        @{
+          type       = "Microsoft.HybridCompute/machines"
+          apiVersion = "2023-03-15-preview"
+          name       = "{{vmName}}"
+          kind       = "VMware"
+          location   = "{{location}}"
+          properties = @{}
         }
-      ]
+        @{
+          type             = "Microsoft.ConnectedVMwarevSphere/VirtualMachineInstances"
+          apiVersion       = "2023-03-01-preview"
+          name             = "default"
+          scope            = "[concat('Microsoft.HybridCompute/machines', '/', '{{vmName}}')]"
+          properties       = @{
+            infrastructureProfile = @{
+              inventoryItemId = "{{vCenterId}}/InventoryItems/{{moRefId}}"
+            }
+          }
+          extendedLocation = @{
+            type = "CustomLocation"
+            name = "{{customLocationId}}"
+          }
+          dependsOn        = @(
+            "[resourceId('Microsoft.HybridCompute/machines','{{vmName}}')]"
+          )
+        }
+      )
     }
   }
 }
-'@
 
 # ARM Template part for Guest Management
-$GMtpl = @'
-{
-  "type": "Microsoft.Resources/deployments",
-  "apiVersion": "2021-04-01",
-  "name": "{{vmName}}-guestmgmt",
-  "dependsOn": [
+$GMtpl = @{
+  type       = "Microsoft.Resources/deployments"
+  apiVersion = "2021-04-01"
+  name       = "{{vmName}}-guestmgmt"
+  dependsOn  = @(
     "[resourceId('Microsoft.Resources/deployments','{{vmName}}-vmcreation')]"
-  ],
-  "properties": {
-    "mode": "Incremental",
-    "template": {
-      "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#",
-      "contentVersion": "1.0.0.0",
-      "resources": [
-        {
-          "type": "Microsoft.HybridCompute/machines",
-          "apiVersion": "2023-03-15-preview",
-          "name": "{{vmName}}",
-          "kind": "VMware",
-          "location": "{{location}}",
-          "properties": {},
-          "identity": {
-            "type": "SystemAssigned"
+  )
+  properties = @{
+    mode     = "Incremental"
+    template = @{
+      '$schema'      = "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#"
+      contentVersion = "1.0.0.0"
+      resources      = @(
+        @{
+          type       = "Microsoft.HybridCompute/machines"
+          apiVersion = "2023-03-15-preview"
+          name       = "{{vmName}}"
+          kind       = "VMware"
+          location   = "{{location}}"
+          properties = @{
           }
-        },
-        {
-          "type": "Microsoft.ConnectedVMwarevSphere/VirtualMachineInstances/guestAgents",
-          "apiVersion": "2023-03-01-preview",
-          "name": "default/default",
-          "scope": "[concat('Microsoft.HybridCompute/machines', '/', '{{vmName}}')]",
-          "properties": {
-            "provisioningAction": "install",
-            "credentials": {
-              "username": "{{username}}",
-              "password": "{{password}}"
-            }
-          },
-          "dependsOn": [
-            "[resourceId('Microsoft.HybridCompute/machines','{{vmName}}')]"
-          ]
+          # identity   = @{
+          #   type = "SystemAssigned"
+          # }
         }
-      ]
+        @{
+          type       = "Microsoft.ConnectedVMwarevSphere/VirtualMachineInstances/guestAgents"
+          apiVersion = "2023-03-01-preview"
+          name       = "default/default"
+          scope      = "[concat('Microsoft.HybridCompute/machines', '/', '{{vmName}}')]"
+          properties = @{
+            provisioningAction = "install"
+            credentials        = @{
+              username = "{{username}}"
+              password = "{{password}}"
+            }
+          }
+          dependsOn  = @(
+            "[resourceId('Microsoft.HybridCompute/machines','{{vmName}}')]"
+          )
+        }
+      )
     }
   }
 }
-'@
 
-$deploymentTemplate = @'
-{
-  "$schema": "https://schema.management.azure.com/schemas/2015-01-01/deploymentTemplate.json#",
-  "contentVersion": "1.0.0.0",
-  "resources": [
-    {{resources}}
-  ]
+$deploymentTemplate = @{
+  '$schema'      = "https://schema.management.azure.com/schemas/2015-01-01/deploymentTemplate.json#"
+  contentVersion = "1.0.0.0"
+  resources      = @()
 }
-'@
 
 #EndRegion: ARM Template
 
@@ -213,9 +219,25 @@ LogText @"
 Starting script with the following parameters:
   VCenterId: $VCenterId
   EnableGuestManagement: $EnableGuestManagement
-  VMCountPerDeployment: $VMCountPerDeployment
-  DryRun: $DryRun
+  VMInventoryFile: $VMInventoryFile
+  VMCredential: $VMCredential
+  Execute: $Execute
 "@
+
+if (!(Test-Path $VMInventoryFile -PathType Leaf)) {
+  LogError "VMInventoryFile not found: $VMInventoryFile"
+  exit
+}
+
+$attemptedVMs = $null
+if ($VMInventoryFile -match "\.csv$") {
+  $attemptedVMs = Import-Csv -Path $VMInventoryFile
+} elseif ($VMInventoryFile -match "\.json$") {
+  $attemptedVMs = Get-Content -Path $VMInventoryFile | ConvertFrom-Json
+} else {
+  LogError "Invalid VMInventoryFile: $VMInventoryFile. Expected file format: CSV or JSON."
+  exit
+}
 
 if (!(Get-Command az -ErrorAction SilentlyContinue)) {
   LogError "az command is not found. Please install azure cli before running this script."
@@ -259,98 +281,136 @@ LogText "Extracted location: $location"
 
 $vmInventoryList = az connectedvmware vcenter inventory-item list --resource-group $resourceGroupName --vcenter $vCenterName --query '[?kind == `VirtualMachine`].{moRefId:moRefId, moName:moName, managedResourceId:managedResourceId}' -o json | ConvertFrom-Json
 
-LogText "Found $($vmInventoryList.Length) VMs in the inventory"
-
-$nonManagedVMs = @()
-
-for ($i = 0; $i -lt $vmInventoryList.Length; $i++) {
-  if (!$vmInventoryList[$i].managedResourceId) {
-    $nonManagedVMs += $vmInventoryList[$i]
+$moRefId2Inv = @{}
+foreach ($vm in $vmInventoryList) {
+  $moRefId2Inv[$vm.moRefId] = @{
+    moName = $vm.moName
+    managedResourceId = $vm.managedResourceId
   }
 }
 
-LogText "Found $($nonManagedVMs.Length) non-managed VMs in the inventory"
+LogText "Found $($attemptedVMs.Length) VMs in the inventory file."
+LogText "Found $($vmInventoryList.Length) VMs in the vCenter inventory, will only enable those which are present in the inventory file."
+
+if ($EnableGuestManagement -and !$VMCredential) {
+  $VMCredential = Get-Credential -Message "Enter the VM credentials for enabling guest management"
+}
 
 function normalizeMoName() {
   param(
     [Parameter(Mandatory=$true)]
     [string]$name
   )
-  return $name.toLower() -replace "[^a-z0-9-]", "-"
-}
-
-$maxVMCountPerDeployment = 200
-if (!$EnableGuestManagement) {
-  $maxVMCountPerDeployment = 400
-}
-
-if (!$VMCountPerDeployment) {
-  $vmCountPerDeployment = $maxVMCountPerDeployment
-}
-else {
-  if ($VMCountPerDeployment -gt $maxVMCountPerDeployment) {
-    LogError "Invalid VMCountPerDeployment: $VMCountPerDeployment. Max allowed value is 400 if guest management is not enabled, else 200. Using max allowed value."
-    $VMCountPerDeployment = $maxVMCountPerDeployment
+  $res = $name -replace "[^A-Za-z0-9-]", "-"
+  if ($res.Length -gt 53) {
+    $res = $res.Substring(0, 52) + "-"
   }
-  $vmCountPerDeployment = $VMCountPerDeployment
+  return $res
 }
+
+$armTemplateLimit = 800
 
 $resources = @()
+$resCountInDeployment = 0
+$batch = 0
 
-for ($i = 0; $i -lt $nonManagedVMs.Length; $i++) {
-  $moRefId = $nonManagedVMs[$i].moRefId
-  
-  # NOTE: Modify the following line to customize the VM name.
-  $vmName = normalizeMoName $nonManagedVMs[$i].moName
-  $vmName += "-$moRefId"
+$summary = @()
 
-  $vmResource = $VMtpl `
-    -replace "{{location}}", $location `
-    -replace "{{vmName}}", $vmName `
-    -replace "{{moRefId}}", $moRefId `
-    -replace "{{vCenterId}}", $VCenterId `
-    -replace "{{customLocationId}}", $customLocationId
-  $resources += $vmResource
+for ($i = 0; $i -lt $attemptedVMs.Length; $i++) {
+  $moRefId = $attemptedVMs[$i].moRefId
+
+  if (!$moRefId2Inv.ContainsKey($moRefId)) {
+    LogDebug "Warning: VM with moRefId $moRefId not found in the vCenter inventory in azure. Skipping."
+    $summary += [ordered]@{
+      vmName     = "$($attemptedVMs[$i].vmName)"
+      moRefId    = $moRefId
+      enabled    = $false
+      guestAgent = $false
+    }
+    continue
+  }
+
+  $resCntCurr = 0
+
+  $inv = $moRefId2Inv[$moRefId]
+
+  $vmName = normalizeMoName $inv.moName
+  $alreadyEnabled = $false
+
+  if ($inv.managedResourceId) {
+    if ($inv.managedResourceId.Contains($VMWARE_RP_NAMESPACE)) {
+      $alreadyEnabled = $true
+    }
+    $resInfo = Get-ARMPartsFromID $inv.managedResourceId
+    $vmName = $resInfo.Name
+  }
+
+  if (!$alreadyEnabled) {
+    $vmResource = $VMtpl | ConvertTo-Json -Depth 30
+    $vmResource = $vmResource `
+      -replace "{{location}}", $location `
+      -replace "{{vmName}}", $vmName `
+      -replace "{{moRefId}}", $moRefId `
+      -replace "{{vCenterId}}", $VCenterId `
+      -replace "{{customLocationId}}", $customLocationId `
+      | ConvertFrom-Json
+    $resCntCurr += 2
+    $resources += $vmResource
+  }
 
   if ($EnableGuestManagement) {
     # NOTE: Set the username and password here. You can also use environment variables to fetch the username and password.
     $username = "Administrator"
     $password = "Password"
 
-    $gmResource = $GMtpl `
+    $gmResource = $GMtpl | ConvertTo-Json -Depth 30
+    $gmResource = $gmResource `
       -replace "{{location}}", $location `
       -replace "{{vmName}}", $vmName `
       -replace "{{username}}", $username `
-      -replace "{{password}}", $password
+      -replace "{{password}}", $password `
+      | ConvertFrom-Json
+
+    if ($alreadyEnabled) {
+      $gmResource.dependsOn = @()
+    }
+    $resCntCurr += 2
     $resources += $gmResource
   }
 
-  $totalBatches = [int](($nonManagedVMs.Length + $vmCountPerDeployment -1) / $vmCountPerDeployment)
-  if (($i + 1) % $vmCountPerDeployment -eq 0 -or ($i + 1) -eq $nonManagedVMs.Length) {
-    $deployment = $deploymentTemplate -replace "{{resources}}", ($resources -join ",")
+  $summary += [ordered]@{
+    vmName = $vmName
+    moRefId = $moRefId
+    enabled = !$alreadyEnabled
+    guestAgent = $true
+  }
 
-    $batch = [int](($i + 1) / $vmCountPerDeployment)
+  if (($resCountInDeployment + 4) -ge $armTemplateLimit -or ($i + 1) -eq $attemptedVMs.Length) {
+    $deployment = $deploymentTemplate | ConvertTo-Json -Depth 30 | ConvertFrom-Json
+    $deployment.resources = $resources
+
+    $batch += 1
     $deploymentName = "vmw-dep-$StartTime-$batch"
     $deploymentFilePath = Join-Path $PSScriptRoot -ChildPath "$deploymentName.json"
 
-    # NOTE: Uncomment the following lines if you want to pretty print the ARM deployment files.
-    # $deployment = ConvertFrom-Json | ConvertTo-Json -Depth 100
-
     $deployment `
+    | ConvertTo-Json -Depth 30
     | Out-File -FilePath $deploymentFilePath -Encoding UTF8
+    if (!$Execute) {
+      $deploymentId = "/subscriptions/$subId/resourceGroups/$resourceGroupName/providers/Microsoft.Resources/deployments/$deploymentName"
+      $deploymentUrl = "https://portal.azure.com/#resource$($deploymentId)/overview"
+      Add-Content -Path $deploymentUrlsFilePath -Value $deploymentUrl
 
-    $deploymentId = "/subscriptions/$subId/resourceGroups/$resourceGroupName/providers/Microsoft.Resources/deployments/$deploymentName"
-    $deploymentUrl = "https://portal.azure.com/#resource$($deploymentId)/overview"
-    Add-Content -Path $deploymentUrlsFilePath -Value $deploymentUrl
+      LogText "(Batch $batch) Deploying $deploymentFilePath"
 
-    LogText "($batch / $totalBatches) Deploying $deploymentFilePath"
-    
-    if (!$DryRun) {
       az deployment group create --resource-group $resourceGroupName --name $deploymentName --template-file $deploymentFilePath --verbose *>> $logFile
     }
     $resources = @()
 
     # NOTE: set sleep time between deployments here, if needed.
+    LogText "Sleeping for 5 seconds before running next batch"
     Start-Sleep -Seconds 5
   }
 }
+
+$summary | ConvertTo-Csv | Out-File vmw-dep-summary.csv
